requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
django-filter==23.3
django-allauth==0.57.0
django-crispy-forms==2.1
Pillow==10.1.0
psycopg2-binary==2.9.9
redis==5.0.1
django-redis==5.3.0
celery==5.3.4
channels==4.0.0
channels-redis==4.1.0
daphne==4.0.0
elasticsearch==8.11.0
python-decouple==3.8
whitenoise==6.6.0
gunicorn==21.2.0